# backend/packages/contracts/openapi/orderflow-gateway-v1.yaml
openapi: 3.0.3
info:
  title: OrderFlow API Gateway
  version: 1.0.0
  description: Публичный HTTP API через api-gateway (префикс /api/v1).
servers:
  - url: http://localhost:3000/api/v1
paths:
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /catalog/products:
    get:
      operationId: catalogListProducts
      tags: [catalog]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /catalog/products/{sku}:
    get:
      operationId: catalogGetProduct
      tags: [catalog]
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /orders:
    get:
      operationId: ordersList
      tags: [orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      operationId: ordersCreate
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >
            Optional. Same key + same body within TTL returns the original order (201 body).
            Same key + different body → 422. Concurrent same key → 409.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderBody'
      responses:
        '201':
          description: Created (or replay of a previous create with the same Idempotency-Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409':
          description: Idempotency-Key already in progress
        '422':
          description: Idempotency-Key reused with a different body
  /orders/{id}:
    get:
      operationId: ordersGet
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}/events:
    get:
      operationId: ordersStatusEvents
      tags: [orders]
      summary: SSE stream of order status changes until terminal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: text/event-stream with JSON payloads { orderId, status, updatedAt }
          content:
            text/event-stream:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthTokens:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    UserProfile:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    Product:
      type: object
      required: [id, sku, name, price, currency, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        currency:
          type: string
        category:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderStatus:
      type: string
      enum: [PENDING, PAYMENT_PENDING, CONFIRMED, CANCELLED, FAILED]
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
    Order:
      type: object
      required: [id, userId, status, totalAmount, currency, createdAt, updatedAt, items]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    CreateOrderItemInput:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
    CreateOrderBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItemInput'
        currency:
          type: string
          minLength: 3
          maxLength: 3